backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
sqlalchemy==2.0.46
aiosqlite==0.22.1
python-jose==3.5.0
python-multipart==0.0.22
pydantic==2.11.7
greenlet==3.3.2
ecdsa==0.19.1
anyio==4.9.0
h11==0.16.0
sniffio==1.3.1
starlette==0.52.1
annotated-types==0.7.0
typing_extensions==4.14.0
